% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{data}
\name{aScatter}
\alias{aScatter}
\title{ascatter R6 Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
aScatter
}
\description{
Create a VR 3D Scatter Plot.
}
\section{Methods}{

\itemize{
  \item{\code{data} Add data.}
  \item{\code{build} Build graph.}
  \item{\code{get} Get graph.}
  \item{\code{browse} Browse graph.}
  \item{\code{embed} Embed graph.}
}
}

\examples{

data(population)

aScatter$
  new(title = "Random")$
  data(
    population, lon, pop, lat, color, size,
   scale = "2 2 2", valfill="1, 9745.6",
   yLimit = 0.2, rotation = "0 90 0"
  )$
  build()$
  embed()

}
\keyword{datasets}
