% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependency.R
\name{as_dependency}
\alias{as_dependency}
\alias{d3_dependency}
\alias{as_full_dependency}
\title{Attach dependency}
\usage{
as_dependency()

d3_dependency()

as_full_dependency()
}
\description{
Attach scatter plot dependency.
}
\note{
The package also depends on d3.
}
\examples{
aframer::a_scene(
  aframer::a_sky(),
  list(
    d3_dependency(),
    as_dependency()
  )
)

# OR
aframer::a_scene(
  aframer::a_sky(),
  as_full_dependency()
)

}
