% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{population}
\alias{population}
\title{Population data}
\format{An object of class \code{data.frame} with 38654 rows and 5 columns.}
\source{
\url{https://ecomfe.github.io/echarts-examples/public/data-gl/asset/data/population.json}
}
\usage{
population
}
\description{
Unreliable, only for demonstration purposes.
}
\keyword{datasets}
